random tour

+++ src/A1/City.java
@@ -3,12 +3,12 @@ package A1;
 import java.util.Random;
 
 public class City {
-    float xCo;
-    float yCo;
+    double xCo;
+    double yCo;
 
     public City() {
         Random rand = new Random();
-        this.xCo = rand.nextFloat();
-        this.yCo = rand.nextFloat();
+        this.xCo = rand.nextDouble();
+        this.yCo = rand.nextDouble();
     }
 }
